+++ CHANGELOG.md
@@ -51,4 +51,5 @@
 - Expose DNS from Consul server one
 - Tidy up Terraform configuration
 - Use Consul servers for DNS
+- Enable script checks
 - Update documentation